packages/pn-personafisica-webapp/src/api/notifications/Notifications.api.ts
import { AxiosResponse } from 'axios';

import {
  GetNotificationsParams,
  GetNotificationsResponse,
  LegalFactId,
  NotificationDetailOtherDocument,
} from '@pagopa-pn/pn-commons';

import { NotificationId } from '../../models/Notifications';
import { DocType, DownloadFileResponse } from '../../redux/notification/types';
import { apiClient } from '../apiClients';
import {
  NOTIFICATIONS_LIST,
  NOTIFICATION_DETAIL_DOCUMENTS,
  NOTIFICATION_DETAIL_LEGALFACT,
  NOTIFICATION_DETAIL_OTHER_DOCUMENTS,
  NOTIFICATION_ID_FROM_QRCODE,
} from './notifications.routes';

const getDownloadUrl = (response: AxiosResponse, docType?: DocType): DownloadFileResponse => {
  if (response.data) {
    return { ...response.data, docType } as DownloadFileResponse;
  }
  return { url: '' } as DownloadFileResponse;
};

export const NotificationsApi = {
  /**
   * Gets current user notifications
   * @param  {string} startDate
   * @param  {string} endDate
   * @returns Promise
   */
  getReceivedNotifications: (
    params: GetNotificationsParams<string>
  ): Promise<GetNotificationsResponse> =>
    apiClient.get<GetNotificationsResponse>(NOTIFICATIONS_LIST(params)).then((response) => {
      if (response.data && response.data.resultsPage) {
        return response.data;
      }
      return {
        resultsPage: [],
        moreResult: false,
        nextPagesKey: [],
      };
    }),

  /**
   * Get notification iun and mandate id from aar link
   * @param {string} qrCode
   * @returns Promise
   */
  exchangeNotificationQrCode: (qrCode: string): Promise<NotificationId> =>
    apiClient
      .post<NotificationId>(NOTIFICATION_ID_FROM_QRCODE(), { aarQrCodeValue: qrCode })
      .then((response) => response.data),

  /**
   * Gets current user notification document aka Atto Notificato
   * @param  {string} iun
   * @param  {number} documentIndex
   * @param  {string} mandateId
   * @returns Promise
   */
  getReceivedNotificationDocument: (
    iun: string,
    documentIndex: string,
    mandateId?: string
  ): Promise<DownloadFileResponse> =>
    apiClient
      .get<DownloadFileResponse>(NOTIFICATION_DETAIL_DOCUMENTS(iun, documentIndex, mandateId))
      .then((response) => getDownloadUrl(response)),

  /**
   * Get notification AAR document
   * @param  {string} iun
   * @param  {NotificationDetailOtherDocument} otherDocument
   * @returns Promise
   */
  getReceivedNotificationOtherDocument: (
    iun: string,
    otherDocument: NotificationDetailOtherDocument,
    mandateId?: string
  ): Promise<DownloadFileResponse> =>
    apiClient
      .get<DownloadFileResponse>(NOTIFICATION_DETAIL_OTHER_DOCUMENTS(iun, otherDocument), {
        params: { documentId: otherDocument.documentId, mandateId },
      })
      .then((response) => getDownloadUrl(response, DocType.AAR)),

  /**
   * Gets current user notification legalfact aka AO3
   * @param  {string} iun
   * @param  {LegalFactId} legalFact
   * @param  {string} mandateId
   * @returns Promise
   */
  getReceivedNotificationLegalfact: (
    iun: string,
    legalFact: LegalFactId,
    mandateId?: string
  ): Promise<DownloadFileResponse> =>
    apiClient
      .get<DownloadFileResponse>(NOTIFICATION_DETAIL_LEGALFACT(iun, legalFact, mandateId))
      .then((response) => getDownloadUrl(response, DocType.AO3)),
};
